Refactor gettrackbyuid to utils file (#55)
* format everything

* Move getTrackByUid to utils file
