+++ src/utils/view-config.ts
@@ -1,4 +1,12 @@
-import { Overlay, ViewConfig } from '../view-config-types';
+import { Overlay, Track, TrackType, ViewConfig } from '../view-config-types';
+
+/**
+ * Should only contain UIDs for "constant" tracks such as combined type tracks.
+ * Tracks with dynamic UIDs based on tileset UIDs will not work in an enum.
+ */
+export const enum CombinedTrackUid {
+  ARCS_AND_BARS = 'arcs-stacked-bars',
+}
 
 export const enum TrackUidPrefix {
   ARCS = 'arcs-track',
@@ -10,6 +18,41 @@ export const enum TrackUidPrefix {
 export const enum TrackOverlayUid {
   REGION_FOCUS = 'region-focus',
   TSS = 'tss',
+}
+
+export function getTrackByUid(
+  viewConfig: ViewConfig,
+  uid: string,
+  byPrefix: boolean = true
+): Track {
+  const topTracks = viewConfig.views[0].tracks.top;
+  if (!topTracks) {
+    throw new Error('No tracks found in top track layout');
+  }
+  const topTracksFlattened = topTracks
+    .map((track) => {
+      if (track.type === TrackType.COMBINED) {
+        return [track, ...track.contents];
+      }
+      return track;
+    })
+    .flat();
+  const trackCandidate = byPrefix
+    ? topTracksFlattened.find((track) => track.uid.startsWith(uid))
+    : topTracksFlattened.find((track) => track.uid === uid);
+  if (!trackCandidate) {
+    throw new Error(`No track found with uid: ${uid}`);
+  }
+  return trackCandidate;
+}
+
+export function replaceTrackByType(
+  trackList: Track[],
+  type: TrackType,
+  newTrack: Track
+) {
+  const index = trackList.findIndex((track) => track.type === type);
+  trackList[index] = newTrack;
 }
 
 export function getOverlayByUid(
