+++ src/view-config-typed.ts
@@ -3,26 +3,14 @@ import { DEFAULT_STRATIFICATION } from './state/stratification-state';
 import {
   DEFAULT_ENHANCER_GENE_ARC_TRACK,
   DEFAULT_ENHANCER_GENE_STACKED_BAR_TRACK,
-  EnhancerGeneTrackInfo,
 } from './state/enhancer-region-state';
 import { DEFAULT_VARIANT_TRACK_DEF } from './state/variant-track-state';
+import { ViewConfig, TrackType } from './view-config-types';
 import {
-  ViewConfig,
-  TrackType,
-  Track,
-  OneDimensionalArcTrack,
-  StackedBarTrack,
-  StratifiedBedTrack,
-} from './view-config-types';
-import { TrackOverlayUid, TrackUidPrefix } from './utils/view-config';
-
-/**
- * Should only contain UIDs for "constant" tracks such as combined type tracks.
- * Tracks with dynamic UIDs based on tileset UIDs will not work in an enum.
- */
-export const enum CombinedTrackUid {
-  ARCS_AND_BARS = 'arcs-stacked-bars',
-}
+  CombinedTrackUid,
+  TrackOverlayUid,
+  TrackUidPrefix,
+} from './utils/view-config';
 
 export const DEFAULT_X_DOMAIN_START = 1761366260;
 export const DEFAULT_X_DOMAIN_END = 1761603836;
@@ -272,38 +260,3 @@ export const DEFAULT_VIEW_CONFIG_ENHANCER: ViewConfig = {
     },
   ],
 };
-
-export function getTrackByUid(
-  viewConfig: ViewConfig,
-  uid: string,
-  byPrefix: boolean = true
-): Track {
-  const topTracks = viewConfig.views[0].tracks.top;
-  if (!topTracks) {
-    throw new Error('No tracks found in top track layout');
-  }
-  const topTracksFlattened = topTracks
-    .map((track) => {
-      if (track.type === TrackType.COMBINED) {
-        return [track, ...track.contents];
-      }
-      return track;
-    })
-    .flat();
-  const trackCandidate = byPrefix
-    ? topTracksFlattened.find((track) => track.uid.startsWith(uid))
-    : topTracksFlattened.find((track) => track.uid === uid);
-  if (!trackCandidate) {
-    throw new Error(`No track found with uid: ${uid}`);
-  }
-  return trackCandidate;
-}
-
-export function replaceTrackByType(
-  trackList: Track[],
-  type: TrackType,
-  newTrack: Track
-) {
-  const index = trackList.findIndex((track) => track.type === type);
-  trackList[index] = newTrack;
-}
